requirements.txt
urllib
pyyaml
pytube
requests
pypresence
youtube_dl
youtube-search
youtubesearchpython